rename nu_test_support::network -> net
