+++ crates/nu-test-support/src/lib.rs
@@ -7,7 +7,7 @@ pub mod fs;
 pub mod harness;
 pub mod locale_override;
 pub mod macros;
-pub mod network;
+pub mod net;
 pub mod playground;
 
 pub mod tester;
