+++ crates/nu-test-support/src/net/proxy.rs
